store/tikv: move CommitDetails and LockDetails to tikv/util (#23854)
