+++ sessionctx/stmtctx/stmtctx.go
@@ -24,6 +24,7 @@ import (
 	"github.com/pingcap/parser"
 	"github.com/pingcap/parser/model"
 	"github.com/pingcap/parser/mysql"
+	"github.com/pingcap/tidb/store/tikv/util"
 	"github.com/pingcap/tidb/util/disk"
 	"github.com/pingcap/tidb/util/execdetails"
 	"github.com/pingcap/tidb/util/memory"
@@ -520,7 +521,7 @@ func (sc *StatementContext) ResetForRetry() {
 
 // MergeExecDetails merges a single region execution details into self, used to print
 // the information in slow query log.
-func (sc *StatementContext) MergeExecDetails(details *execdetails.ExecDetails, commitDetails *execdetails.CommitDetails) {
+func (sc *StatementContext) MergeExecDetails(details *execdetails.ExecDetails, commitDetails *util.CommitDetails) {
 	sc.mu.Lock()
 	defer sc.mu.Unlock()
 	if details != nil {
@@ -559,7 +560,7 @@ func (sc *StatementContext) MergeTimeDetail(timeDetail execdetails.TimeDetail) {
 }
 
 // MergeLockKeysExecDetails merges lock keys execution details into self.
-func (sc *StatementContext) MergeLockKeysExecDetails(lockKeys *execdetails.LockKeysDetails) {
+func (sc *StatementContext) MergeLockKeysExecDetails(lockKeys *util.LockKeysDetails) {
 	sc.mu.Lock()
 	if sc.mu.execDetails.LockKeysDetail == nil {
 		sc.mu.execDetails.LockKeysDetail = lockKeys
